roleplay-platform/rpg_platform/requirements.txt
djangorestframework==3.14.0
drf-yasg==1.21.7
python-memcached==1.59
django-cors-headers==4.3.1 